+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Bean Searcher
   * `DefaultParamResolver` 新增 `gexprName` 属性 
   * 新增条件分组 与 组逻辑运算功能（TODO）
+  * 新增 `TimeFieldConvertor`，支持 `java.sql.Time` 与 `LocalTime` 之间的转换 
 * Bean Searcher Boot Starter
   * 新增 `bean-searcher.params.gexpr` 配置键，可在配置文件中指组表达式参数名